+++ src/main/java/com/skycat/mystical/mixin/EggEntityMixin.java
@@ -30,18 +30,18 @@ public abstract class EggEntityMixin extends ThrownItemEntity {
     public Entity mystical_onEggCollision(Entity chicken) {
         if (Mystical.getSpellHandler().isConsequenceActive(MysteryEggsConsequence.class) &&
         !Mystical.getHavenManager().isInHaven(chicken) &&
-        Utils.percentChance(Mystical.CONFIG.mysteryEggs.chance())) { // TODO: Config chance
+        Utils.percentChance(Mystical.CONFIG.mysteryEggs.chance())) {
             EntityType<?> randomType = Utils.getRandomEntryFromTag(Registries.ENTITY_TYPE, MysticalTags.RANDOM_EGG_SPAWNABLE);
             if (randomType != null) {
                 Entity randomEntity = randomType.create(chicken.getWorld());
                 if (randomEntity != null) {
                     randomEntity.refreshPositionAndAngles(getX(), getY(), getZ(), getYaw(), 0);
                     chicken.discard();
-                    Utils.log(MysteryEggsConsequence.FACTORY.getFiredMessage()); // TODO: Config loglevel
+                    Utils.log(MysteryEggsConsequence.FACTORY.getFiredMessage(), Mystical.CONFIG.mysteryEggs.logLevel());
                     return randomEntity;
                 }
             }
-            Utils.log("Unable to spawn a random egg mob."); // TODO: Translate
+            Utils.log(Utils.translateString("text.mystical.consequence.mysteryEggs.failedSpawn"));
         }
         return chicken;
     }