+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - Changed some advancement names
 - Fixed nighttime events happening at the wrong times
 - Fixed levitation spell not respecting havens
+- Added mysteryEggs spell